:truck: Rename cards module
